Enable tcltk option in the default windows installation for Auto Build Marlin to work correctly
Please platformio team consider enabling tcltk option for windows installation by default. The Auto Build Marlin plugin for visual studio requires tcl/tk to be installed and has platformio as a dependency. If no python is installed before installing Auto Build Marlin and PlatformIO, the tcl/tk will be missing and the dialog windows of Auto Build Marlin will not work. The user cannot know the reason for that. I have documented the issue already but it would be best if the default installation includes tcl/tk by default to avoid user confusion and frustration.